requirements_minimal.txt
# Core Data Science Libraries
numpy==1.24.4
pandas==2.0.3
scipy==1.10.1

# Visualization
matplotlib==3.7.5
seaborn==0.12.2
plotly==5.18.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# Explainability
shap==0.44.0

# Time Series
prophet==1.1.5
statsmodels==0.14.1

# Dashboard
streamlit==1.29.0

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
